Fixed missing versionlens.onShowingProgress command
