+++ src/presentation/extension.ts
@@ -1,9 +1,13 @@
 export * from './extension/versionLensExtension';
-export * from './extension/versionLensCommands';
+export * from './extension/commands/suggestionCommands';
 export * from './extension/versionLensState';
 
 export * from './extension/events/textEditorEvents';
 export * from './extension/events/textDocumentEvents';
 
 export * from './extension/models/contextState';
+
 export * from './extension/helpers/installedStatusHelpers';
+export * as CommandHelpers from './extension/helpers/commandHelpers';
+
+export * from './extension/commands/iconCommands';
