+++ src/presentation/extension/commands/suggestionCommands.ts
@@ -1,0 +1,68 @@
+// vscode references
+import * as VsCodeTypes from 'vscode';
+
+import { ILogger } from 'core/logging';
+import { PackageSourceTypes } from 'core/packages';
+
+import { VersionLensExtension } from "../versionLensExtension";
+import { VersionLensState } from '../versionLensState';
+import { VersionLens } from 'presentation/lenses/versionLens';
+import { CommandHelpers } from 'presentation/extension';
+
+export enum SuggestionCommandContributions {
+  UpdateDependencyCommand = 'versionlens.onUpdateDependencyCommand',
+  LinkCommand = "versionlens.onLinkCommand"
+}
+
+export class SuggestionCommands {
+
+  state: VersionLensState;
+
+  extension: VersionLensExtension;
+
+  constructor(extension: VersionLensExtension) {
+    this.extension = extension
+    this.state = extension.state;
+  }
+
+  onUpdateDependencyCommand(codeLens: VersionLens, packageVersion: string) {
+    if ((<any>codeLens).__replaced) return Promise.resolve();
+
+    const { workspace, WorkspaceEdit } = require('vscode');
+    const edit = new WorkspaceEdit();
+    edit.replace(codeLens.documentUrl, codeLens.replaceRange, packageVersion);
+
+    return workspace.applyEdit(edit)
+      .then(done => (<any>codeLens).__replaced = true);
+  }
+
+  onLinkCommand(codeLens: VersionLens) {
+    const path = require('path');
+    const opener = require('opener');
+
+    if (codeLens.package.source === PackageSourceTypes.Directory) {
+      const filePathToOpen = path.resolve(
+        path.dirname(codeLens.documentUrl.fsPath),
+        codeLens.package.resolved.version
+      );
+      opener(filePathToOpen);
+      return;
+    }
+    // opener(codeLens.package.remoteUrl);
+  }
+
+}
+
+export function registerSuggestionCommands(
+  extension: VersionLensExtension, logger: ILogger
+): Array<VsCodeTypes.Disposable> {
+
+  const suggestionCommands = new SuggestionCommands(extension);
+
+  return CommandHelpers.registerCommands(
+    SuggestionCommandContributions,
+    <any>suggestionCommands,
+    logger
+  )
+
+}
