+++ src/presentation/lenses.ts
@@ -1,4 +1,4 @@
 export * from './lenses/definitions/iVersionCodeLens';
-export * as CommandFactory from './lenses/commands/commandFactory';
+export * as CommandFactory from './lenses/commands/suggestionCommandFactory';
 export * as VersionLensFactory from './lenses/versionLensFactory';
 export * from './lenses/versionLens';
